requirements.txt
Appium-Python-Client==2.11.1
behave==1.2.6
selenium==4.11.2
urllib3==1.26.6
allure-behave==2.13.2
#google
gspread==5.11.2
oauth2client==4.1.3
allure-combine==1.0.11
#Test
pytest==7.4.0
